+++ src/os/unix.rs
@@ -12,7 +12,7 @@ pub trait VirtualInput<T: Send> where Self: Sized {
     fn create_virtual<F>(
         self, port_name: &str, callback: F, data: T
     ) -> Result<MidiInputConnection<T>, ConnectError<Self>>
-    where F: FnMut(f64, &[u8], &mut T) + Send + 'static;
+    where F: FnMut(u64, &[u8], &mut T) + Send + 'static;
 }
 
 /// Trait that is implemented by `MidiOutput` on platforms that
